Make server's socket writing and client's socket reading dynamic. Fixes #56.
Former-commit-id: aa90938a72fb73c8935e97a6e64281d7aa07a4ea

+++ server/src/include/newclient.h
@@ -351,17 +351,17 @@ typedef struct CS_Stats
 /** Srv client files. */
 typedef struct _srv_client_files
 {
-	/** File data, compressed or not */
+	/** Compressed file data. */
 	char *file;
 
-	/** If -1, the file is not compressed */
-	int len;
+	/** Compressed file length. */
+	size_t len;
 
 	/** Original uncompressed file length */
-	int len_ucomp;
+	size_t len_ucomp;
 
-	/** CRC adler32 */
-	unsigned int crc;
+	/** CRC32 sum. */
+	unsigned long crc;
 } _srv_client_files;
 
 /** Srv client files. */
